+++ packages/manager/apps/pci-databases-analytics/src/__tests__/helpers/mocks/userAccess.ts
@@ -1,0 +1,6 @@
+import * as database from '@/types/cloud/project/database';
+
+export const mockedKafkaUserAccess: database.kafka.user.Access = {
+  cert: 'certificates',
+  key: 'keyCertificates',
+};