+++ packages/manager/apps/pci-databases-analytics/src/pages/services/[serviceId]/users/resetPassword/ResetPassword.modal.tsx
@@ -155,7 +155,7 @@ const ResetUserPassword = () => {
                 mode="outline"
                 data-testid="reset-password-close-button"
               >
-                {t('resetUserPasswordButtonClose')}
+                {t('userButtonClose')}
               </Button>
             </DialogClose>
           ) : (