+++ packages/manager/apps/pci-instances/src/data/hooks/instance/useInstance.ts
@@ -26,6 +26,7 @@ export const useInstance = ({
   const projectId = useProjectId();
 
   return useQuery({
+    ...queryOptions,
     queryKey: instancesQueryKey(projectId, [
       'region',
       String(region),
@@ -34,6 +35,5 @@ export const useInstance = ({
       ...(params ?? []),
     ]),
     queryFn: () => getInstance({ projectId, region, instanceId, params }),
-    ...queryOptions,
   });
 };