+++ packages/manager/apps/pci-instances/src/pages/instances/action/components/modal/ActionModalContent.component.tsx
@@ -88,6 +88,7 @@ export const ActionModalContent: FC<TActionModalProps> = ({
 
   return isLoading ? (
     <div className="pt-6">
+      {/* eslint-disable-next-line @typescript-eslint/no-unsafe-assignment  */}
       {[...new Array(3)].map((_elt, index) => (
         <OsdsSkeleton key={index} />
       ))}