+++ packages/manager/apps/pci-kubernetes/src/pages/list/Mode.tsx
@@ -3,7 +3,9 @@ import { useRegionInformations } from '@/api/hooks/useRegionInformations';
 
 const Mode = ({ projectId, region }: { projectId: string; region: string }) => {
   const { data } = useRegionInformations(projectId ?? '', region);
-  return <>{data && <RegionChipByType type={data.type} showTooltip />}</>;
+  return (
+    <>{data && <RegionChipByType type={data.type} showTooltip={false} />}</>
+  );
 };
 
 export default Mode;