+++ packages/manager/apps/pci-databases-analytics/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.26.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-databases-analytics-app@1.25.2...@ovh-ux/manager-pci-databases-analytics-app@1.26.0) (2025-08-12)
+
+
+### Bug Fixes
+
+* **pci-db:** metrics are not well displayed for mongodb engine ([80d18ce](https://github.com/ovh/manager/commit/80d18cec519d4e6987215046868594cb1626c38b)), closes [#DATATR-2426](https://github.com/ovh/manager/issues/DATATR-2426)
+
+
+### Features
+
+* **pci-databases:** add certificates management for kafka user ([6dada6e](https://github.com/ovh/manager/commit/6dada6ec74d6248233df57c4a04ffd21e69dca76)), closes [#DATATR-2416](https://github.com/ovh/manager/issues/DATATR-2416)
+
+
+
+
+
 ## [1.25.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-databases-analytics-app@1.25.1...@ovh-ux/manager-pci-databases-analytics-app@1.25.2) (2025-08-07)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-databases-analytics-app